+++ packages/cpp-exceptions-test/src/catch.cpp
@@ -49,6 +49,16 @@ catch_call_pyobj(PyObject* x)
   return msg;
 }
 
+extern "C" void
+set_suspender(__externref_t suspender);
+
+extern "C" char*
+promising_catch_call_pyobj(__externref_t suspender, PyObject* x)
+{
+  set_suspender(suspender);
+  return catch_call_pyobj(x);
+}
+
 jmp_buf my_jump_buffer;
 void
 longjmp_func(int status);